requirements.txt
fastapi==0.111.0
mysql_connector_repackaged==0.3.1
